\name{ages}
\alias{ages}

\title{
ages
}
\description{
 A function to compute age from ID number.
}
\usage{
ages(x)
}

\arguments{
  \item{x}{
   a numeric or vector
}

}
\details{

}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%%
}
\author{
 DataXujing
}
\note{

}

%% Make other sections like Warning with \section{Warning }{....}


\examples{
# not run

#-------------ages Example-----------------------

library(stringr)
ages(x)
#obtain a numeric or vector about ages

}

