\name{genders}
\alias{genders}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  genders
}
\description{
   A function to compute gender from ID number.
}
\usage{
genders(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
     a numeric or vector like ages()
}

 

}
\details{
%%  Non
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% Non
}
\author{
  DataXujing
}
\note{
%%  Non
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{

# not run 

#-----genders Example-------------------------------
library(stringr)
genders(x)
# obtain a numeric or vector about genders
}
