\name{idprep-package}
\alias{idprep-package}
\alias{idprep}
\docType{package}
\title{
\packageTitle{idprep}
}
\description{
\packageDescription{idprep}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{idprep}
\packageIndices{idprep}
 A packages to transmit ID number.
}
\author{
\packageAuthor{idprep}

Maintainer: \packageMaintainer{idprep}
}
\references{
  


}
\examples{

# not run 

library(stringr)

#-----ages Example---------------------------------

ages(x)

#-----genders Example-------------------------------

genders(x)

#-------------vvply Example-----------------------


vvply(X,FUN,...)

#-------------Province Example-----------------------


Province(X)

#------------- div   Example-----------------------


div(X,Y)


}
